feat: 이름 도메인 생성
