+++ README.md
@@ -45,16 +45,16 @@ graph TD
 - [ ] 카드를 받는다.
 - [ ] 점수를 확인한다.
 - [ ] 카드 추가 여부를 결정한다.
+- [ ] 여러 명일 수 있다.
+  - [ ] 중복되는 이름은 가질 수 없다.
+  - [ ] 최대 5명이다.
 
 ### 플레이어
 
-- [ ] 이름을 가진다.
+- [x] 이름을 가진다.
   - [x] 최소 1자, 최대 10자까지 가능하다.
-  - [ ] 중간 공백은 허용한다.
-  - [ ] 중복되는 이름은 가질 수 없다.
+  - [x] 중간 공백은 허용한다.
   - [x] `딜러`라는 이름은 가질 수 없다.
-- [ ] 여러 명일 수 있다.
-  - [ ] 최대 5명이다.
 
 ### 카드
